GLFW, GLAD and GLM dependencies added and linked to engine project GLM linked to test project added engine dll to ignore list when it's in the _WorkingDir

+++ Projects/MallardEngine/Application.h
@@ -1,5 +1,7 @@
 #pragma once
 
+
+
 #include "DLLBuild.h"
 
 class Application {
@@ -7,6 +9,6 @@ public:
 	DLL_BUILD Application();
 	DLL_BUILD ~Application();
 
-	DLL_BUILD int getRandom();
+	DLL_BUILD float getRandom();
 };
